Restructure App and move to DPLP
* Restructured app again, making it flat. UI, STD and ENC daemons are part of the same artifact. This makes access modifiers useful too.
* Changed from OSPFv2 to DPLP (for IPv4 only?).
* Added protocol data types from specification
*  Changed routerID from ipv4address to short (16 bit). May need to change from short, as short is 2's C. All Java natives are.
* Removed OSPF data types.

DPLP is my own protocol draft specification from COM740. It sparked the idea of encrypting a routing protocol, which this research is attempting to create. a 'standard' version was not specified originally, so sensible changes to the specification need to be made to make it not encrypted. For example, choosing to require challenge-response still, or using unencrypted hello solicitation message type for all hello messages.

Apparently Artifact is spelt "Artefact" in the UK, which isn't my or my Nan's intuition.