Fixes in Run 1 for Test 7
Added Javadocs to UI

Added region comments to UI

Changed Launcher.operationMode to package-private, to allow package classes to change behaviour based on std daemon or enc daemon

Changed the on dead timer expire method for NeighbourNodes. Nodes are no longer removed from the neighbours list. Their states are only set to Down. Also called SendHelloPacket for the current daemon in use, as neighbour down is a topology change, and so all neighbours should be updated of this change.

Removed assigning known neighbours in first packet handler, as it is added anyway below for all neighbour nodes

Removed ability to specify a single interface to send hello packets over. It will never be used, as quick event updates are actually topological changes, and so all nodes need to know them as soon as they happen.

Fixed MakeHelloPacket logic to only append neighbours not in the ExternalStates.Down state

Fixed MakeHelloPacket logic for specifying packet length, by moving the append to inside the for each neighbour loop. Here, buffers are not needed to store the neighbours list as a byte array, and the result of multiple concats rather than one single concat

Fixed Node GetKnownNeighboursString exception when the known neighbours list was empty